Déplace la page api-gestionnaire dans module partner-api
